+++ R/convention.R
@@ -1,19 +1,21 @@
 #' Zarr convention
 #'
-#' @description This class implements a basic Zarr convention. A convention is a
-#'   set of attributes specific to a certain domain of application. These
-#'   attributes are included in Zarr group and array attributes and are
-#'   interpreted by application code. Conventions may be grouped in domains and
-#'   combined with other conventions.
+#' @description This class implements a basic Zarr convention attribute factory.
+#'   A convention is a set of attributes specific to a certain domain of
+#'   application. These attributes are included in Zarr group and array
+#'   attributes and are interpreted by application code. Conventions may be
+#'   grouped in domains and combined with other conventions.
 #'
 #'   Application-specific conventions need to inherit from this base class and
 #'   redefine relevant methods. Descendant conventions may have additional
-#'   methods specific to the domain of the data.
+#'   methods specific to the domain of the data. Descendants of this class take
+#'   the elements that define it and then return them formatted for inclusion in
+#'   the attributes of a Zarr node.
 #'
 #'   It is not useful to directly instantiate this class, use a descendant
 #'   convention instead. It is recommended that descendant classes use the
-#'   "zarr_conv_****" naming pattern and that they are included in a R package
-#'   with similar conventions and/or domain(s).
+#'   "zarr_convention_*" naming pattern and that they are included in a R
+#'   package with similar conventions and/or domain(s).
 #' @docType class
 #' @export
 zarr_convention <- R6::R6Class('zarr_convention',
@@ -59,14 +61,26 @@ zarr_convention <- R6::R6Class('zarr_convention',
       attributes
     },
 
-    #' @description Write the data of a convention instance in the attributes of
-    #'   a Zarr object. This method does not do any actual writing. Descendant
-    #'   classes should implement their specific solutions.
-    #' @param attributes A `list` with Zarr attributes for a group or array. The
-    #'   properties will be written at the root level of `attributes`.
-    #' @return The updated attributes.
-    write = function(attributes) {
-      attributes
+    #' @description Set the attributes for this convention for use in a Zarr
+    #'   node. This is a stub that descendant classes can implement, using a
+    #'   specific set of arguments. More complex conventions can use other
+    #'   arrangements to set the more complex attributes.
+    set = function() {
+      # No-op
+    },
+
+    #' @description Format the elements of a convention instance in a list
+    #'   suitable for the attributes of a Zarr object. Descendant classes should
+    #'   implement their specific solutions.
+    #' @return The convention attributes in a list.
+    as_list = function() {
+      list()
+    },
+
+    #' @description Clear any attributes that may have been set. Only the
+    #' properties of the convention itself will remain in place.
+    clear = function() {
+      # No-op
     }
   ),
   active = list(