+++ NEWS.md
@@ -1,7 +1,9 @@
 # zarr (development version)
 
+- `zarr_node::post_open()` method allows for processing that requires the Zarr hierarchy to be in place.
 - Dynamically set a node in a Zarr hierarchy.
 - Set metadata on a node in a memory store.
+- Convention classes are now coded as attribute factories.
 - Fix key listing in memory stores.
 
 # zarr 0.4.1